+++ example/ios/Runner/RuntimeStub.m
@@ -34,6 +34,9 @@
     self = [super init];
     if (self) {
         _object = [NSObject new];
+        dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(1 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
+            [NSNotificationCenter.defaultCenter postNotificationName:@"SampleDartNotification" object:nil];
+        });
     }
     return self;
 }