+++ example/lib/delegatestub.dart
@@ -1,12 +1,21 @@
 import 'package:dart_objc/dart_objc.dart';
 
-class DelegateStub extends NSObject {
+abstract class SampleDelegate {
+  callback();
+}
+
+class DelegateStub extends NSObject implements SampleDelegate {
   DelegateStub() : super(Class('DelegateStub', type(of: NSObject))) {
-    registerCallback(callback, 'callback', 'StubDelegate');
+    registerProtocolCallback(callback, 'callback', 'StubDelegate');
+    registerNotificationCallback(handleNotification, 'handleNotification:');
   }
 
   callback() {
     print('callback succeed!');
     return NSObject();
+  }
+
+  handleNotification(NSObject notification) {
+    print('receive notification!');
   }
 }